+++ RouteHandlerInterface.php
@@ -2,7 +2,9 @@
 
 namespace LoremIpsum\RouteGeneratorBundle;
 
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+
 interface RouteHandlerInterface
 {
-    public function handle($value, $view = null, $context = []);
+    public function handle($value, $view = null, $context = [], $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH);
 }