+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # RouteGeneratorBundle Changelog
 
+## [0.2.0] - 2019-04-23
+### Add
+- Support route $referenceType (defaults to `UrlGeneratorInterface::ABSOLUTE_PATH`)
+
+### Change
+- **BK-CHANGE** Change interface RouteHandlerInterface and RouteGeneratorInterface to support $referenceType
+
 ## [0.1.3] - 2018-12-27
 ### Fix
 - Fix possible string conversion error when using routable